Return, payment, cancel methods added.

+++ src/Bike.java
@@ -1,6 +1,6 @@
 public class Bike extends Vehicle {
 
-    public Bike(int id, String brand, double pricePerDay, double kilometersDriven) {
+    public Bike(int id, String brand, double pricePerDay, int kilometersDriven) {
         super(id, brand, pricePerDay, kilometersDriven);
     }
